feat: app refactored and styled
