+++ src/components/layouts/header.js
@@ -5,7 +5,7 @@ import React from "react";
 function Header() {
   return (
     <header className="header">
-      <h1>Is Your 🎂 Birthday Lucky? 🤔</h1>
+      <h1>Is Your Birthday Lucky? 🎂 🤔</h1>
     </header>
   );
 }
